fix reST highlighting of bare code blocks
Move the field detection rule to the very end so that the rule for
detecting bare code blocks (i.e. "::" on a line by itself) fires first.
This is simpler than tweaking the already-complicated field detection
regexpr to require a non-empty field name, and is okay now that role
detection is more strict.