fix highlighting of reST block comments
Add highlighting support for block comments in reST, modeled after the
code block support. As this makes the comment rule fire sooner than many
other '..' rules, also refactor all of those to have '^\s*' at the
beginning of their patterns, and while we're at it, add 'column="0"' to
every rule that has a beginning-of-line assertion.