Auto-adapt hard-coded colors to current color scheme.
This code is used like that for years in KDevelop and works quite
nicely for most cases.

Note that for darker color schemes, this most often results in
a bit less-intense colors which is not a big problem as colors stand
out much more against a dark background. And its better than before,
where you'd end up with hard-coded colors for a bright background
scheme...

Bright color schemes are not affected, as long as they fulfill the
conditition of having a lower luma for the background color compared
to the foreground color.

REVIEW: 115928