Highlight Qt4 extensions by default.
The user chose the Qt4 highlighter for a reason. If he doesn't want
to highlight these things, he should use the standalone C++
highlighter.

Classes are highlighted as dsOthers as is commonly used by other
highlighting files. Functions use dsFunction. The macros use the
old style which we had before.

Furthermore the file is cleaned up by removing 'Qt4 Extensions' -
these extensions are just macros as well.

Additionally all extensions that are valid for the default C++
highlighter are added to the Qt4 highlighter as well.