Remove "AlignedCommentends" feature from isocpp.xml.
This feature is broken and very confusing to people. We had a
discussion on this and most people want to have ti removed. Just
now again a user was complaining about it:

> Could somebody tell me how blue "//" comment symbols are
> different from their gray counterparts?

Since this feature depends on the indent-depth and also requires
indentation with spaces, it should not be shipped with the default
syntax highlight, which must cope with all kinds of indentation
configurations.

CCMAIL: kwrite-devel@kde.org