Do not reset the model without need
From reading the code i see no need to reset the model, the only think that i see the reset do is cause the view to "reset" too and the state of open/closed roots in the tree gets broken.

Used it for a few minutes and everything worked fine. Reading the code don't see why it shouldn't. I'll it use it a bit more for a few more days.

BUGS: 332990