Fix crash/assert on opening files with similar paths
In KateFileTreeModel::handleDuplicitRootDisplay we need to check if check_root has been removed in the loop of rdir, if it has there's no need to remove it on the nrdir branch, doing so asserts on debug builds and on release builds it creates a broken tree structure that will crash later when trying to use it.

Doesn't crash anymore, added new test that passes, old tests still pass.

Acked by Michal Humpula and Christoph Cullmann

BUGS: 347578
REVIEW: 123793