XML Compltion for Kate Syntax Files: Add new default styles
Since we dropped language.dtd in favour of language.xsd, it's not possible
to auto-generate this file anymore. That is quite bad, since we need to
maintain the meta DTD file manually now... This is OK for small changes as
this one, but it'll probably break as soon as we do larger changes to
language.xsd... Not exactly cool.

CHANGELOG: Honor new default styles in Kate XML completion plugin.