Make it possible to open files at a certain line or column.
That was possible before, but now we support the much more common
format of file:line:column. All of that is optional of course.
You can also open just file:line.

Note that this check is only done for local files, not remote ones.

Dominik said he'll post-review this patch once I pushed it.