fix wrong KMoreTools API usage
BUG: 356615

The previously used method
KMoreToolsMenuFactory.fillMenuFromGroupingNames returns a lazy menu
which is not filled until the menu is shown. But in the following code
we do not show the menu but use its menu items to fill another menu. But
there were never any menu items in gitMenu->actions().

Solution: use KMoreToolsMenuFactory.fillMenuFromGroupingNames instead.