Search & Replace Plugin: Set foreground color for matches to default foreground text color
This improves the readability a lot, since the background color now does
not have to have good contrast with ALL foreground colors in the color scheme.
Instead, it only needs to have good contrast in only one case.
This also helps for dark color schemes.

This is the application addition for review 127554.
CHANGELOG: Improved foreground color contrast for search & replace highlightings