explicitly initialize kcrash
Summary:
To quote the documentation
> Call this in your main() to ensure that the crash handler is always
> launched.

The problem in particular is that cmake will not actually link kcrash even
if it is in the link list but not used. ::initialize for the most part just
makes sure that we actually link against kcrash and thus have it
initialized via qt static init magic.

Reviewers: dhaumann

Reviewed By: dhaumann

Subscribers: dhaumann, kwrite-devel, #kate

Differential Revision: https://phabricator.kde.org/D2460