Fix location of Open action in the toolbar.
The file_open standard action should always appear after file_new. This
is not the case in the Kate toolbar, where we have Open before New.

This happens because there is a custom file_open group that breaks the
kxmlgui merging index (the file_open action gets 0 rather than 1).

This was introduced a long time ago with:

    https://websvn.kde.org/?view=revision&revision=350820

The log message just says

    "to allow the template plugin a button in a relevant position"

which is no longer relevant (the plugin was dropped in fbb7bb4285).
For the record, see also a03c76f3e9.

Dropping this group makes Open appear in the toolbar after New, as expected.

Differential Revision: https://phabricator.kde.org/D8142