Search&Replace: trim file types list from whitespace
When splitting the string lists for the filetypes to search
and the files to exclude, strip leading and trailing whitespace
from the list members. My colleague stumbled a few times upon this,
when the search found nothing because of a space before the next
matching pattern.

CCMAIL:kare.sars@iki.fi

Alex