Use Tree View as the default view mode in the Filesystem Browser plugin
Summary:
Previously, Kate's Filesystem Browser plugin was replying on a bug in KDirOperator that was just fixed: the fact that columns perviously did not auto-size correctly and did not stick to the right edge when the view was resized. This was fixed in D11993.

Now that this bug is fixed, the plugin needs to default to a more sensible view mode--one that does not have additional columns. This patch does that.

Test Plan:
{nav Create and log into a new user account > open Kate > Settings > Configure Kate > Plugins > turn on "Filesystem Browser" plugin > hit Apply > click "Filesystem Browser" on the sidebar}

Before:
{F5808728}

After:
{F5808725}

Reviewers: #kate, dhaumann

Reviewed By: #kate, dhaumann

Subscribers: dhaumann, sharvey

Tags: #kate

Differential Revision: https://phabricator.kde.org/D12224