Downloading syntax highlighting was removed in KTextEditor 5.50
Downloading syntax highlighting files was removed in the
KTextEditor Framework version 5.50. The reason for this is
that the KTextEditor framework is released once a month, meaning
that users who use latest KDE software usually anways have the
most recent version. Other users who do not follow the latest
development releases are likely not the target audience for
downloading highlighting files.

Besides that, there are technical problems with only updating
certain highlighting files, since it can lead to an inconsistent
state when one highlighting file needs another one that was not
updated or does not exist, or also if a highlighting file needs
a certain indenter that does not yet exist or contains bugs.

In addition, we have a nice small cleanup in the UI, since we
have now one button less.