two terminal windows, one for directory with .kateproject and one for the base directory of the project, if different
this allows e.g. to nicely jump between build and source dir for CMake
generated .kateproject files

CHANGELOG: Two terminal toolviews for each project, one for the
directory with the .kateproject file, one for the base directory, if
that is a different directory. This allows to navigate between source
and build directory for e.g. CMake generated .kateproject files