Make "Save Session" action ask for a name for anonymous sessions
Summary:
When invoking "Save Session", kate should ask for a session name if
the session is anonymous, i.e. the same behaviour when using saving a
a new file/unnamed document.

BUG: 319396
FIXED-IN: 19.04.0

Test Plan:
Opening a kate instance with a new/anonymous session and invoking
"Save Session" should ask the user to enter a name for the session.

Reviewers: #kate, cullmann, dhaumann

Reviewed By: #kate, cullmann

Subscribers: ngraham, kwrite-devel, #kate

Tags: #kate

Differential Revision: https://phabricator.kde.org/D19017