Rudimentary xml/html parser for symbolviewer plugin
Summary:
Added a single new source file xml_parser.cpp based on xslt_parser.cpp to produce a
symbol view for xml and html files. Updated existing symbolviewer files. This stretches the
concept of symbolviewer a bit (what is a function or a macro in html), but it is useful for
some editing tasks.

Reviewers: sars

Subscribers: sars, kwrite-devel

Tags: #kate

Differential Revision: https://phabricator.kde.org/D21024