Snippets: Make snippet shortcuts work also if the snippets tool view is minimized, by adding the QActions to the main window instead of the tool view
Summary: The keyboard shortcut for snippets did not work while the snippets tool view was minimized. This seems to be because the snippet QActions must be added to a visible(?) widget to be considered. Thus, this change makes the snippets plugin add its actions to the main window instead of its tool view such that they remain active when the snippets tool view is minimized to the side.

Test Plan: I tested it manually. Note for testing: There seems to be another bug that sometimes prevents the shortcuts for new snippets from working (a restart of the application should make them work). I think that this is unrelated to this change and will try to address it next.

Reviewers: #kate, cullmann

Reviewed By: #kate, cullmann

Subscribers: cullmann, kwrite-devel

Tags: #kate

Differential Revision: https://phabricator.kde.org/D22308