Snippets: Fix the keyboard shortcuts for new snippets not working
Summary: After creating a new snippet with an assigned keyboard shortcut, this shortcut did not work directly. This was because the snippet QActions are added to a widget by a rowsInserted() signal, but the new snippet's QAction did not exist yet at this time. This change ensures that the new snippet's QAction is created in time.

Test Plan: I tested it manually.

Reviewers: #kate, cullmann

Reviewed By: #kate, cullmann

Subscribers: cullmann, kwrite-devel

Tags: #kate

Differential Revision: https://phabricator.kde.org/D22309