When loading a session always load the "open recent" list
Summary:
KateMainWindow: move loading/saving "open recent" files list to separate
functions to enable loading that list when loading a session regardless
of the "save window configuration" setting.

Note that there was no strict need to move the code that saved the list
outside of KateMainWindow::saveProperties(), but this way the code is less
confusing.

BUG: 408499
FIXED-IN: 19.08.0

Test Plan:
- Open kate with a new/anonymous session, and disable "restore window configuration"
- Open some files to populate the "open recent" menu, then close kate
- Open kate again with a new/anon. session; check if the "open recent" menu
  has been populated  with the previously opened files
- Load a different session in the same window; check if the "open recent" menu
  has been populated with the files list from that session

- Repeat the same steps but with "restore window configuration" enabled

Reviewers: #kate, cullmann, dhaumann

Reviewed By: #kate, cullmann

Subscribers: cfeck, kwrite-devel

Tags: #kate

Differential Revision: https://phabricator.kde.org/D21876