LCPClientView: Fix assert with Qt::UniqueConnection
From the Qt docs: If you pass the Qt::UniqueConnection type, the
connection will only be made if it is not a duplicate. If there
is already a duplicate (exact same signal to the exact same slot
on the same objects), the connection will fail and connect will
return an invalid QMetaObject::Connection.

In this case, the first connect worked, but the second time the
connect was already there, so an invalid QMetaObject::Connection
was returned, leading to the assert.

I think we don't need the assert.