Enlarge the find and replace add-on combo boxes
Summary:
When the find and replace toolview is in the left or right sidebars, move the combo boxes on their own rows so they have more space.
This patch changes the position of the combo boxes when the toolview width is smaller than its height.
The goal is to be able to see a large string in the find and/or replace combo boxes within a moderately narrow toolview.

Test Plan:
Original suggestion, widgets are not grouped:
{F6930229}
{F6930232}

2nd suggestion, buttons are small:
{F6931264}
{F6931265}

3rd suggestion:
{F6933433}

toolview height Vs combobox width:

before, 22 result lines, the combobox needs to be expanded
{F7021390}

after, 20 result lines, the combobox is large
{F7021407}

Reviewers: #kate, sars, #vdg, ngraham

Reviewed By: #kate, sars, #vdg, ngraham

Subscribers: cullmann, ngraham, kwrite-devel

Tags: #kate, #vdg

Differential Revision: https://phabricator.kde.org/D22059