Clean up Kate addons CMake scripts.
Summary:
This is the first in a series of revisions modernizing the Kate CMake scripts.
I'm submitting the changes to the addons directory first as they form a
nicely separated list of changes.

Note that if https://phabricator.kde.org/D22698 and
https://phabricator.kde.org/D22699 are merged, most if not all of the
remaining source lists in the addons directory can be removed in favor of
working with targets.

List of changes made:

- Standardize style on two spaces and no whitespace between commands and
  arguments.
- Remove directory commands in favor of target-based commands as
  recommended by modern CMake.
- Remove usage of `qt5_add_resources` in favor of `CMAKE_AUTORCC` and
  adding .qrc files directly to a target's sources.
- Remove usage of source lists (where possible) in favor of CMake 3.1's
  `target_sources` command.
- Remove the CMake binary directory as an include directory from most
  addons.
- Remove unnecessary comments, CMake project calls, config files, etc
  from addon CMake scripts.
- Remove `HAVE_CTERMID` from config.h and move the check and
  addition  to compile definitions to the project addon CMake script as its only
  used in the project plugin sources.

Test Plan: Apply changes and verify Kate still builds and works exactly as before.

Reviewers: #kate, cullmann

Reviewed By: #kate, cullmann

Subscribers: cullmann, kossebau, turbov, asturmlechner, kwrite-devel

Tags: #kate

Differential Revision: https://phabricator.kde.org/D22721