Further cleanup addons CMake scripts.
- Update CMake minimum required version to CMake 3.1 which is when
`target_sources` was added.
- Use `target_sources` even when commands that work on
sources variables are used. Instead of appending the generated sources
to an existing sources variable, we store them in a temporary variable
and immediately add them to the corresponding target with
`target_sources`.
- Remove unnecessary dependencies on `kdeinit_kate`.
- Remove the intermediary katefiletree static library and have the tests
build the single required source file manually instead which is
consistent with how the other addons with tests work.