Add Tools > External Tools > Configure...
This is based on the newly exposed showPluginConfigPage() in 5.63.
But the External Tools plugin is built only if 5.63 is around, anyways.

We have to use a Qt::QueuedConnection, since invoking showConfigPage()
now allows the user to disable the plugin as well, which would delete
the instances of the current code path being executed.

The QueuedConnection works around this, albeit still ugly...
Better fixes welcome.