Search for KF5I18n and KF5DocTools
Summary:
KF5I18n is required anyway. If it isn't searched and found when the
ki18n_install call, which is injected by the release scripts, is
encountered the build will fail with "CMake Error: No script
specified for argument -P" because a variable usually set by
ki18n_install is not defined.
The same applies to KF5DocTools and kdoctools_install as well, but
it is guarded by a "if (KF5DocTools_FOUND)" so it doesn't error out
but just doesn't install any handbooks, even if KF5DocTools is found.

Test Plan: The 19.11.80 tarball builds fine with the patch applied

Reviewers: #kate, cullmann

Subscribers: kwrite-devel

Tags: #kate

Differential Revision: https://phabricator.kde.org/D25411