Read the user visible name of the default color scheme
The user visible name and the name of a colorscheme as configured in kdeglobals
can differ. For example "Breeze Light" and "BreezeLight". This causes no color
scheme to be checked when such a scheme is the default scheme and no color
scheme has been selected in Kate.


(cherry picked from commit ad2e9867a1deb776189ce34356e7861e848add4a)