Add sortable columns to session manager dialog, and remove previous sorting code
Summary:
I really like the new session manager dialog (session chooser) but when I went to click on the column headers to sort, it didn't work. So I wondered how hard it would be to implement and just went for it.

I know the diff is supposed to be done with arc, but I don't yet have that all set up, and I wanted to get you the patch over as soon as possible. I hope that's ok.

One thing that does not work properly is the sorting by number of open files, since that requires numeric sort which was beyond my current abilities, maybe someone else can take that up for the next round, once this simple patch is in.

I also corrected a couple spelling mistakes, which accounts for the extra lines.

Test Plan:
Apply patch (should apply cleanly to most recent master).
Compile.
Run kate.

Reviewers: #kate, dhaumann, cullmann

Reviewed By: #kate, cullmann

Subscribers: kwrite-devel

Tags: #kate

Differential Revision: https://phabricator.kde.org/D27180