Avoid some unneeded clones of snippet repositories.
When an upstream repository is changed, the snippets plugin makes a copy in the personal data dir.
Unfortunately, changes are often detected without any real modification. Importantly, a "rename" is
often triggered simply by clicking on a repository name.

This patch disregards such spurious "changes", and also clarifies the corresponding message a bit.