Show the line text in search results
This is a proposal.

Currently search & replace plugin shows "Line: xx, Column: xx"
which IMO is not very useful. With this change the actual line
is displayed with the matching text highlighted in the same color
as that of the editor search-highlight color.

This makes the search plugin very useful for cases where you want
to see some very similar lines or see uses of something. An example
use case which I often come across is that suppose you have a big log
file and you are looking for only WARNs. You just type WARN and in
search results you see all the lines that contain WARN. Now you can
quickly analyze them.