[ColorPicker] Use QVector instead of QHash for storing the indices
QHash::keys() returns a QList which needs to be converted into a QVector
for the return value, resulting to 2 allocations on a very hot code path.

The color indices are now stored in a pair of QVectors, one for the
color note indices (that needs to be returned in inlineNotes()) and
the other holds the corresponding indices. This removes the wasteful
allocations, but slightly increases memory consumption.