Improve search performance by reducing the amount of work
This change takes advantage of my previous change and reuses the
stored colors instead of creating them from scratch every time.

rootFileItem() now returns early for search-as-you-type and CurrentFile
searches. This greatly reduces the amount of work it has to do.

The same file which took 9-10 seconds in search now takes 5 seconds only