[ColorPicker] Fix Color note updating logic
Previously, it was assumed that changes are always contiguous and that
no change in the number of lines means that only one line needs to be updated.
These assumptions do not cover e.g. multiline search and replace and
multiline modification (without new lines being added).

For simplicity in implementing the changes, the end parameter of
ColorPickerInlineNoteProvider::updateNotes() is now exclusive.