Allow / separated pattern to filter same filename across different directories
Suppose you have the following dir structure:

abc/main.cpp
def/main.cpp

Using quickopen it is very difficult to find the main.cpp that you want.
This is because we remove the filename from the path so the filter
doesn't work if you try for e.g "ab/ma".

For now, I have made it so that the filename is always present in the
path but is removed during display. This allows for more flexible
filtering.