Tweak fuzzy matcher to score sequential matches a bit higher for kate command bar
Currently if you search for "Sort", the result for "sort" is 2nd or
third and not the first because Camel Case / separator matches have a
higher score. To fix this sequential score is raised a bit as a special
case for Kate command bar by introducing a new
`fuzzy_match_sequential()` function.