Use the new theme function to provide better semantic highlighting
This change is a second proposal for better semantic highlighting for
C++. It can be considered as temporary till we are able to support for
new API and till LSPs themselves add proper support for offical semantic
tokens.

Since we rely on the new API call (theme()), this change will only work
for KF > 5.79.

BUG: 431014
BUG: 431011