LSP: Fix Ctrl + click underline highlighting broken in #include context
Current in a #include <some/file.h> if you ctrl + click, you get
underline under "some" or "file" and not even "file.h" which makes it
look very weird and as if "some" is something else and "file.h" is
something else, when infact they form one thing and point to one thing.
This change fixes this by expanding the range to include everything
between <> or "".