Only match filePath when filename matched OR pattern has '/'
If filename didn't match, it is highly unlikely that filePath will get
matched. However, if there is a '/' in the pattern, we will match the
path anyways because then we are sure the user is trying to match the
path.

The main reason this is necessary is because matching path is really
slow because paths are long so we try to avoid matching path as much as
we can.