Allow users to enable indexing by clicking button
Before there was the instruction to enable indexing by modifying
some config file without mentioning the format/where it is located.
Also the instruction did not seem to work, because the value is read
from ~/.config/katerc.

This will make it far easier to enable indexing. The setting
in the projectsplugin config works too, but that is not as discoverable.